Override default rails jQuery ujs functionality to provide double click deletion confirmation.

+++ app/helpers/posts_helper.rb
@@ -9,7 +9,7 @@ module PostsHelper
 
   def display_like_or_unlike_button(likable)
     if current_user.likes? (likable)
-      link_to "Unlike", likes_path(user_id: current_user.id, likable_id: likable.id, likable_type: likable.class), method: 'delete', remote: true
+      link_to "Unlike", likes_path(user_id: current_user.id, likable_id: likable.id, likable_type: likable.class), method: 'delete', data: { confirm: "Sure?" }, remote: true
     else
       link_to "Like", likes_path(likable_id: likable.id, likable_type: likable.class), method: 'post', remote: true
     end
